Rename search_tool.rs to search.rs
