+++ src/tools/mod.rs
@@ -3,7 +3,7 @@
 pub mod bash;
 pub mod edit_file;
 pub mod sandbox;
-pub mod search_tool;
+pub mod search;
 pub mod skill;
 pub mod write_file;
